# docs-backup-20260213-150302/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ATO Tax Optimizer API
  version: 0.1.0
  description: >
    API for the Australian Tax Optimizer platform. Integrates with Xero for
    financial data, uses Gemini AI for forensic analysis, and provides tax
    optimisation recommendations for Australian businesses.
  contact:
    name: ATO Tax Optimizer
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://ato-app.vercel.app
    description: Production (Vercel)
security:
  - bearerAuth: []
...
paths:
...
  /api/accountant/verify:
    get:
      tags: [accountant]
      summary: Verify accountant credentials
      description: Verifies professional credentials with issuing body.
      parameters:
        - name: credentialNumber
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Verification result
...
  /api/accountant/findings/{id}/status:
    patch:
      tags: [accountant]
      summary: Update finding status
      description: Updates the review status of an accountant finding.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
      responses:
        '200':
          description: Status updated
...
  /api/accountant/reports/generate:
    post:
      tags: [accountant]
      summary: Generate accountant report
      description: Generates a professional report for accountant distribution.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId]
              properties:
                tenantId:
                  type: string
      responses:
        '200':
          description: Report generated

  # ── Alerts ──
...
  /api/alerts/{id}:
    patch:
      tags: [alerts]
      summary: Update alert
      description: Updates alert status (dismiss, snooze, etc.).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert updated
    delete:
      tags: [alerts]
      summary: Delete alert
      description: Permanently deletes an alert.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from authentication
...